Add files via upload
This project aims to automate the testing of a banking application using TestNG and Page Factory design pattern. The application under test is a web-based banking system accessible through the URL https://qaebank.ccbp.tech/ebank/login. The testing will cover the functionality and user interface of both the Login page and the Home page.